rust-checks: Move the meat of checks to library crate
This will allow us to use it directly from the main crate.
