+++ rust-checks/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod analysis;
+pub mod checks;
 pub mod comment_finders;
 pub mod common_structs;
 pub mod tree_utils;
